refac: move things around, uplift oauth endpoints
